Removing config attribute from TrainingCallbaclAttributes
