--- nerfstudio/configs/trainer_config.py
@@ -1,56 +1,0 @@
-# Copyright 2022 The Nerfstudio Team. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""Config used for running an experiment"""
-
-from __future__ import annotations
-
-from dataclasses import dataclass, field
-from pathlib import Path
-from typing import Optional, Type
-
-from rich.console import Console
-
-from nerfstudio.configs.experiment_config import ExperimentConfig
-from nerfstudio.engine.trainer import Trainer
-
-CONSOLE = Console(width=120)
-
-
-@dataclass
-class TrainerConfig(ExperimentConfig):
-    """Configuration for training regimen"""
-
-    _target: Type = field(default_factory=lambda: Trainer)
-    """target class to instantiate"""
-    steps_per_save: int = 1000
-    """Number of steps between saves."""
-    steps_per_eval_batch: int = 500
-    """Number of steps between randomly sampled batches of rays."""
-    steps_per_eval_image: int = 500
-    """Number of steps between single eval images."""
-    steps_per_eval_all_images: int = 25000
-    """Number of steps between eval all images."""
-    max_num_iterations: int = 1000000
-    """Maximum number of iterations to run."""
-    mixed_precision: bool = False
-    """Whether or not to use mixed precision for training."""
-    save_only_latest_checkpoint: bool = True
-    """Whether to only save the latest checkpoint or all checkpoints."""
-    # optional parameters if we want to resume training
-    load_dir: Optional[Path] = None
-    """Optionally specify a pre-trained model directory to load from."""
-    load_step: Optional[int] = None
-    """Optionally specify model step to load from; if none, will find most recent model in load_dir."""
-    load_config: Optional[Path] = None
