+++ tests/test_train.py
@@ -10,9 +10,9 @@ from pathlib import Path
 import pytest
 
 from nerfstudio.configs.method_configs import method_configs
-from nerfstudio.configs.trainer_config import TrainerConfig
 from nerfstudio.data.dataparsers.blender_dataparser import BlenderDataParserConfig
 from nerfstudio.data.dataparsers.minimal_dataparser import MinimalDataParserConfig
+from nerfstudio.engine.trainer import TrainerConfig
 from scripts.train import train_loop
 
 BLACKLIST = ["base", "semantic-nerfw", "instant-ngp", "instant-ngp-bounded", "nerfacto", "phototourism"]
